requirements.txt
# Copyright 2014 PerfKitBenchmarker Authors. All rights reserved.
#
# Licensed under the Apache License, Version 2.0 (the "License");
# you may not use this file except in compliance with the License.
# You may obtain a copy of the License at
#
#   http://www.apache.org/licenses/LICENSE-2.0
#
# Unless required by applicable law or agreed to in writing, software
# distributed under the License is distributed on an "AS IS" BASIS,
# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
# See the License for the specific language governing permissions and
# limitations under the License.
click==8.0.3
tabulate
loguru>=0.5.0
rsa>=4.7.2
absl-py
jinja2>=2.10.2
setuptools>=40.3.0
colorlog[windows]==2.6.0
blinker>=1.3
packaging
pandas>=1.1.5
PyYAML>=5.4.1
pint
numpy>=1.16.5,<2.0
contextlib2>=0.5.1
six>=1.13.0
pywinrm
timeout-decorator
google-cloud-aiplatform
google-cloud-core
google-cloud-datastore
google-cloud-monitoring>=2.0.0
beautifulsoup4
requests
seaborn
ptyprocess==0.7.0
python-dotenv==1.0.1

# unlike setup.py requirements.txt cannot require a Python version AFAICT
# this hack requires a non-existent package when Python is less than 3.11
PerfkitBenchmarker_requires_Python_311;python_version<"3.11"
